Deduplicate e2e helpers and tidy party-polish structure.
Extract shared Playwright helpers, collapse ProgressPanel waiting copy into a phase helper, and inline zero-logic reveal string wrappers.
