+++ client/e2e/full-round.e2e.ts
@@ -1,4 +1,5 @@
-import { expect, test, type Browser, type BrowserContext, type Page } from '@playwright/test';
+import { expect, test, type BrowserContext, type Page } from '@playwright/test';
+import { createPlayers, drawStroke, makeAppUrl } from './helpers';
 
 interface TestPlayer {
   name: string;
@@ -18,7 +19,9 @@ test('one TV and three phones complete a full drawing round', async ({ baseURL, 
     await expect(tv.locator('.room-code')).toHaveText(/[A-Z]{4}/);
     const roomCode = (await tv.locator('.room-code').innerText()).trim();
 
-    const players = await createPlayers(browser, contexts, appUrl, roomCode, ['Ava', 'Bo', 'Cy']);
+    const names = ['Ava', 'Bo', 'Cy'];
+    const pages = await createPlayers(browser, contexts, appUrl, roomCode, names);
+    const players: TestPlayer[] = pages.map((page, index) => ({ name: names[index], page }));
     await expect(tv.locator('.player-row')).toHaveCount(players.length);
     for (const player of players) {
       await expect(tv.getByText(player.name, { exact: true })).toBeVisible();
@@ -81,74 +84,6 @@ test('one TV and three phones complete a full drawing round', async ({ baseURL, 
     await Promise.all(contexts.map((context) => context.close()));
   }
 });
-
-function makeAppUrl(baseURL: string | undefined): (path: string) => string {
-  if (!baseURL) {
-    throw new Error('Playwright baseURL is required for Draw Party e2e tests.');
-  }
-  return (path: string) => new URL(path, baseURL).toString();
-}
-
-async function createPlayers(
-  browser: Browser,
-  contexts: BrowserContext[],
-  appUrl: (path: string) => string,
-  roomCode: string,
-  names: string[]
-): Promise<TestPlayer[]> {
-  const players: TestPlayer[] = [];
-  for (const name of names) {
-    const context = await browser.newContext({
-      hasTouch: true,
-      isMobile: true,
-      viewport: { width: 390, height: 844 }
-    });
-    contexts.push(context);
-
-    const page = await context.newPage();
-    await page.goto(appUrl(`/join/${roomCode}`));
-    await expect(page.locator('input.code-input')).toHaveValue(roomCode);
-    await page.getByPlaceholder('Your name').fill(name);
-    await page.getByRole('button', { name: 'Join the Party' }).click();
-    await expect(page.locator('.app-shell.player .brand')).toHaveText('Lobby');
-    await expect(page.getByText(`${name}, you're in`)).toBeVisible();
-    players.push({ name, page });
-  }
-  return players;
-}
-
-async function drawStroke(page: Page): Promise<void> {
-  const canvas = page.locator('canvas.draw-canvas');
-  await expect(canvas).toBeVisible();
-  await expect
-    .poll(async () => {
-      const box = await canvas.boundingBox();
-      return Boolean(box && box.width >= 100 && box.height >= 75);
-    })
-    .toBe(true);
-
-  await canvas.evaluate((element: HTMLCanvasElement) => {
-    const rect = element.getBoundingClientRect();
-    const fire = (type: string, xRatio: number, yRatio: number, buttons = 1) => {
-      element.dispatchEvent(
-        new PointerEvent(type, {
-          bubbles: true,
-          cancelable: true,
-          pointerId: 1,
-          pointerType: 'pen',
-          isPrimary: true,
-          buttons,
-          clientX: rect.left + rect.width * xRatio,
-          clientY: rect.top + rect.height * yRatio
-        })
-      );
-    };
-    fire('pointerdown', 0.2, 0.25);
-    fire('pointermove', 0.45, 0.45);
-    fire('pointermove', 0.7, 0.3);
-    fire('pointerup', 0.7, 0.3, 0);
-  });
-}
 
 async function waitForArtistIndex(players: TestPlayer[]): Promise<number> {
   let currentArtistIndex = -1;
