+++ client/e2e/reliability.e2e.ts
@@ -1,4 +1,5 @@
-import { expect, test, type Browser, type BrowserContext, type Page } from '@playwright/test';
+import { expect, test, type BrowserContext } from '@playwright/test';
+import { createPlayers, drawStroke, makeAppUrl, waitForGuessers } from './helpers';
 
 test('late join mid-drawing becomes spectator and promotes next round', async ({ baseURL, browser }) => {
   const contexts: BrowserContext[] = [];
@@ -101,98 +102,3 @@ test('disconnected drawer does not stall drawing phase', async ({ baseURL, brows
     await Promise.all(contexts.map((context) => context.close().catch(() => undefined)));
   }
 });
-
-function makeAppUrl(baseURL: string | undefined): (path: string) => string {
-  if (!baseURL) {
-    throw new Error('Playwright baseURL is required for Draw Party e2e tests.');
-  }
-  return (path: string) => new URL(path, baseURL).toString();
-}
-
-async function createPlayers(
-  browser: Browser,
-  contexts: BrowserContext[],
-  appUrl: (path: string) => string,
-  roomCode: string,
-  names: string[]
-): Promise<Page[]> {
-  const players: Page[] = [];
-  for (const name of names) {
-    const context = await browser.newContext({
-      hasTouch: true,
-      isMobile: true,
-      viewport: { width: 390, height: 844 }
-    });
-    contexts.push(context);
-    const page = await context.newPage();
-    await page.goto(appUrl(`/join/${roomCode}`));
-    await page.getByPlaceholder('Your name').fill(name);
-    await page.getByRole('button', { name: 'Join the Party' }).click();
-    await expect(page.getByText(`${name}, you're in`)).toBeVisible();
-    players.push(page);
-  }
-  return players;
-}
-
-async function drawStroke(page: Page): Promise<void> {
-  const canvas = page.locator('canvas.draw-canvas');
-  await expect(canvas).toBeVisible();
-  const box = await canvas.boundingBox();
-  if (!box) {
-    throw new Error('Missing canvas box');
-  }
-  await page.evaluate(
-    ({ x, y, width, height }) => {
-      const target = document.querySelector('canvas.draw-canvas');
-      if (!(target instanceof HTMLCanvasElement)) {
-        throw new Error('Missing draw canvas');
-      }
-      const startX = x + width * 0.25;
-      const startY = y + height * 0.35;
-      const endX = x + width * 0.7;
-      const endY = y + height * 0.65;
-      for (const [type, clientX, clientY] of [
-        ['pointerdown', startX, startY],
-        ['pointermove', endX, endY],
-        ['pointerup', endX, endY]
-      ] as const) {
-        target.dispatchEvent(
-          new PointerEvent(type, {
-            bubbles: true,
-            cancelable: true,
-            pointerId: 1,
-            pointerType: 'touch',
-            isPrimary: true,
-            clientX,
-            clientY,
-            buttons: type === 'pointerup' ? 0 : 1
-          })
-        );
-      }
-    },
-    box
-  );
-}
-
-async function waitForGuessers(players: Page[]): Promise<Page[]> {
-  const expected = Math.max(0, players.length - 1);
-  await expect
-    .poll(async () => {
-      let count = 0;
-      for (const player of players) {
-        if (await player.getByPlaceholder('Something that sounds legit…').isVisible().catch(() => false)) {
-          count += 1;
-        }
-      }
-      return count;
-    })
-    .toBe(expected);
-
-  const found: Page[] = [];
-  for (const player of players) {
-    if (await player.getByPlaceholder('Something that sounds legit…').isVisible().catch(() => false)) {
-      found.push(player);
-    }
-  }
-  return found;
-}
